+++ pyogrio/tests/test_raw_io.py
@@ -173,14 +173,16 @@ def test_read_bbox_invalid(naturalearth_lowres, bbox):
         read(naturalearth_lowres, bbox=bbox)
 
 
-def test_read_bbox(naturalearth_lowres):
+def test_read_bbox(naturalearth_lowres_all_ext):
     # should return no features
     with pytest.warns(UserWarning, match="does not have any features to read"):
-        geometry, fields = read(naturalearth_lowres, bbox=(0, 0, 0.00001, 0.00001))[2:]
+        geometry, fields = read(
+            naturalearth_lowres_all_ext, bbox=(0, 0, 0.00001, 0.00001)
+        )[2:]
 
     assert len(geometry) == 0
 
-    geometry, fields = read(naturalearth_lowres, bbox=(-140, 20, -100, 40))[2:]
+    geometry, fields = read(naturalearth_lowres_all_ext, bbox=(-140, 20, -100, 45))[2:]
 
     assert len(geometry) == 2
     assert np.array_equal(fields[3], ["USA", "MEX"])
@@ -243,7 +245,7 @@ def test_read_fids_unsupported_keywords(naturalearth_lowres):
         read(naturalearth_lowres, fids=[1], where="iso_a3 = 'CAN'")
 
     with pytest.raises(ValueError, match="cannot set both 'fids' and any of"):
-        read(naturalearth_lowres, fids=[1], bbox=(-140, 20, -100, 40))
+        read(naturalearth_lowres, fids=[1], bbox=(-140, 20, -100, 45))
 
     with pytest.raises(ValueError, match="cannot set both 'fids' and any of"):
         read(naturalearth_lowres, fids=[1], skip_features=5)