+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 -   register GDAL drivers during initial import of pyogrio (#145)
 -   support writing "not a time" (NaT) values in a datetime column (#146)
+-   fixes an error when reading GPKG with bbox filter (#150) 
+-   properly raises error when invalid where clause is used on a GPKG (#150)
 -   avoid duplicate count of available features (#151)
 
 ## 0.4.1